Guia 2 Java POO Ejercicio 03

+++ Guia2_POO_ej03/src/guia2_poo_ej03/servicios/ServiciosOperacion.java
@@ -63,7 +63,7 @@ public class ServiciosOperacion {
         }
     }
     
-    public double dividir(){
+    public float dividir(){
         if (num1==0 || num2==0){
             System.out.println("Dividendo o Divisor es 0");   
         return 0;
